Humanize labels in the Ukrainian language (#278)
This set of edits corrects some labels in the Ukrainian language for them to sound more natural to a native speaker.

The change also aims to ensure the consistency of the app-specific language (e.g., "revision" to be consistently translated
as "версія", "revert" to be consistently translated as "відновити", and "summary" as "резюме").

Also, some past tense verbs and adjectives are corrected to be gender-neutral. (Arguably, there are better ways to do it,
e.g., via restructuring the sentence to avoid the need for a gender-accorded form, but it is acceptable to just list both masculine&feminine
forms like this "переглянув/переглянула" or like this "переглянув(-ла)").